requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.2.1
email-validator==2.0.0
twilio==8.4.0
python-dotenv==1.0.0
bootstrap-flask==2.3.0
pandas==2.1.0
matplotlib==3.8.0
seaborn==0.13.0
openpyxl==3.1.2
xlsxwriter==3.1.2
weasyprint==60.1
